+++ border-atari-env/examples/random_pong.rs
@@ -3,7 +3,9 @@ use border_atari_env::{
     BorderAtariAct, BorderAtariActRawFilter, BorderAtariEnv, BorderAtariEnvConfig, BorderAtariObs,
     BorderAtariObsRawFilter,
 };
-use border_core::{Configurable, DefaultEvaluator, Env as _, Evaluator, Policy};
+use border_core::{
+    Agent, Configurable, DefaultEvaluator, Env as _, Evaluator, NullReplayBuffer, Policy,
+};
 use serde::Deserialize;
 
 type Obs = BorderAtariObs;
@@ -27,6 +29,8 @@ impl Policy<Env> for RandomPolicy {
         fastrand::u8(..self.n_acts as u8).into()
     }
 }
+
+impl Agent<Env, NullReplayBuffer> for RandomPolicy {}
 
 impl Configurable for RandomPolicy {
     type Config = RandomPolicyConfig;
@@ -60,16 +64,12 @@ fn main() -> Result<()> {
                 n_acts: n_acts as _,
             }
         };
-        RandomPolicy::build(policy_config)
+        Box::new(RandomPolicy::build(policy_config)) as _
     };
 
     // Runs evaluation
     let env_config = env_config.render(true);
-    let _ = {
-        let env = Env::build(&env_config, 0)?;
-        DefaultEvaluator::new(env, 5)?
-    }
-    .evaluate(&mut policy);
+    let _ = DefaultEvaluator::new(&env_config, 42, 5)?.evaluate(&mut policy);
 
     Ok(())
 }