+++ border-core/src/base/agent.rs
@@ -7,13 +7,19 @@ use std::path::Path;
 /// Represents a trainable policy on an environment.
 pub trait Agent<E: Env, R: ReplayBufferBase>: Policy<E> {
     /// Set the policy to training mode.
-    fn train(&mut self);
+    fn train(&mut self) {
+        unimplemented!();
+    }
 
     /// Set the policy to evaluation mode.
-    fn eval(&mut self);
+    fn eval(&mut self) {
+        unimplemented!();
+    }
 
     /// Return if it is in training mode.
-    fn is_train(&self) -> bool;
+    fn is_train(&self) -> bool {
+        unimplemented!();
+    }
 
     /// Performs an optimization step.
     ///
@@ -24,14 +30,33 @@ pub trait Agent<E: Env, R: ReplayBufferBase>: Policy<E> {
     }
 
     /// Performs an optimization step and returns some information.
-    fn opt_with_record(&mut self, buffer: &mut R) -> Record;
+    #[allow(unused_variables)]
+    fn opt_with_record(&mut self, buffer: &mut R) -> Record {
+        unimplemented!();
+    }
 
     /// Save the parameters of the agent in the given directory.
     /// This method commonly creates a number of files consisting the agent
     /// in the directory. For example, the DQN agent in `border_tch_agent` crate saves
     /// two Q-networks corresponding to the original and target networks.
-    fn save_params<T: AsRef<Path>>(&self, path: T) -> Result<()>;
+    #[allow(unused_variables)]
+    fn save_params(&self, path: &Path) -> Result<()> {
+        unimplemented!();
+    }
 
     /// Load the parameters of the agent from the given directory.
-    fn load_params<T: AsRef<Path>>(&mut self, path: T) -> Result<()>;
+    #[allow(unused_variables)]
+    fn load_params(&mut self, path: &Path) -> Result<()> {
+        unimplemented!();
+    }
+
+    #[allow(missing_docs)]
+    fn as_any_ref(&self) -> &dyn std::any::Any {
+        unimplemented!("as_any_ref() must be implemented for train_async()");
+    }
+
+    #[allow(missing_docs)]
+    fn as_any_mut(&mut self) -> &mut dyn std::any::Any {
+        unimplemented!("as_any_mut() must be implemented for train_async()");
+    }
 }