+++ border-core/src/evaluator/default_evaluator.rs
@@ -1,97 +1,51 @@
 use super::Evaluator;
-use crate::{Env, Policy};
+use crate::{record::Record, Agent, Env, ReplayBufferBase};
 use anyhow::Result;
 
 /// A default [`Evaluator`].
 ///
-/// This evaluator runs a given number of episodes and returns the mean value of cumulative reward.
-/// The code for this method is as follows:
-///
-/// ```no_run
-/// # use anyhow::Result;
-/// # use border_core::{Env, Policy, Evaluator};
-/// # use std::marker::PhantomData;
-/// # pub struct DefaultEvaluator<E: Env, P: Policy<E>> {
-/// # n_episodes: usize,
-/// # env: E,
-/// # phantom: PhantomData<P>,
-/// # }
-///
-/// # impl<E, P> DefaultEvaluator<E, P>
-/// # where
-/// # E: Env,
-/// # P: Policy<E>,
-/// # {
-/// # fn run_episode(&mut self, policy: &mut P, init_obs: E::Obs) -> Result<f32> {
-/// # unimplemented!();
-/// # }
-/// # }
-/// # impl<E, P> Evaluator<E, P> for DefaultEvaluator<E, P>
-/// # where
-/// # E: Env,
-/// # P: Policy<E>,
-/// # {
-/// # fn evaluate(&mut self, policy: &mut P) -> Result<f32> {
-/// let mut r_total = 0f32;
-///
-/// for ix in 0..self.n_episodes {
-///    let init_obs = self.env.reset_with_index(ix)?;
-///    r_total += self.run_episode(policy, init_obs)?;
-/// }
-///
-/// Ok(r_total / self.n_episodes as f32)
-/// # }
-/// # }
-/// ```
+/// This struct runs episodes of the given number of times.
 pub struct DefaultEvaluator<E: Env> {
     n_episodes: usize,
     env: E,
 }
 
-impl<E> Evaluator<E> for DefaultEvaluator<E>
-where
-    E: Env,
-{
-    fn evaluate<P: Policy<E>>(&mut self, policy: &mut P) -> Result<f32> {
+impl<E: Env> Evaluator<E> for DefaultEvaluator<E> {
+    fn evaluate<R>(&mut self, policy: &mut Box<dyn Agent<E, R>>) -> Result<Record>
+    where
+        R: ReplayBufferBase,
+    {
         let mut r_total = 0f32;
 
         for ix in 0..self.n_episodes {
-            let init_obs = self.env.reset_with_index(ix)?;
-            r_total += self.run_episode(policy, init_obs)?;
+            let mut prev_obs = self.env.reset_with_index(ix)?;
+
+            loop {
+                let act = policy.sample(&prev_obs);
+                let (step, _) = self.env.step(&act);
+                r_total += step.reward[0];
+                if step.is_done() {
+                    break;
+                }
+                prev_obs = step.obs;
+            }
         }
 
-        Ok(r_total / self.n_episodes as f32)
+        let name = format!("Average return over {} episodes", self.n_episodes);
+        Ok(Record::from_scalar(name, r_total / self.n_episodes as f32))
     }
 }
 
-impl<E> DefaultEvaluator<E>
-where
-    E: Env,
-{
+impl<E: Env> DefaultEvaluator<E> {
     /// Constructs [`DefaultEvaluator`].
     ///
     /// `env` - Instance of the environment.
     /// `n_episodes` - The number of episodes for evaluation.
     ///   The evaluator returns the mean value of cumulative reward in each episode.
-    pub fn new(env: E, n_episodes: usize) -> Result<Self> {
-        Ok(Self { n_episodes, env })
-    }
-
-    /// Runs an episode and returns the cumulative reward.
-    fn run_episode<P: Policy<E>>(&mut self, policy: &mut P, init_obs: E::Obs) -> Result<f32> {
-        let mut r_total = 0f32;
-        let mut prev_obs = init_obs;
-
-        loop {
-            let act = policy.sample(&prev_obs);
-            let (step, _) = self.env.step(&act);
-            r_total += step.reward[0];
-            if step.is_done() {
-                break;
-            }
-            prev_obs = step.obs;
-        }
-
-        Ok(r_total)
+    pub fn new(config: &E::Config, seed: i64, n_episodes: usize) -> Result<Self> {
+        Ok(Self {
+            n_episodes,
+            env: E::build(config, seed)?,
+        })
     }
 }