+++ border-py-gym-env/examples/random_cartpole.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use border_core::{
-    record::Record, Configurable, DefaultEvaluator, Env as _, Evaluator as _, Policy,
+    record::Record, Agent, Configurable, DefaultEvaluator, Evaluator as _, NullReplayBuffer, Policy,
 };
 use border_py_gym_env::{
     ArrayObsFilter, DiscreteActFilter, GymActFilter, GymEnv, GymEnvConfig, GymObsFilter,
@@ -11,7 +11,7 @@ use std::convert::TryFrom;
 type PyObsDtype = f32;
 
 mod obs {
-    use ndarray::{ArrayD, IxDyn};
+    use ndarray::ArrayD;
 
     #[derive(Clone, Debug)]
     pub struct CartPoleObs(ArrayD<f32>);
@@ -56,7 +56,7 @@ type Act = CartPoleAct;
 type ObsFilter = ArrayObsFilter<PyObsDtype, f32, Obs>;
 type ActFilter = DiscreteActFilter<Act>;
 type Env = GymEnv<Obs, Act, ObsFilter, ActFilter>;
-type Evaluator = DefaultEvaluator<Env, RandomPolicy>;
+type Evaluator = DefaultEvaluator<Env>;
 
 #[derive(Clone, Deserialize)]
 struct RandomPolicyConfig;
@@ -77,6 +77,8 @@ impl Configurable for RandomPolicy {
         Self
     }
 }
+
+impl Agent<Env, NullReplayBuffer> for RandomPolicy {}
 
 #[derive(Debug, Serialize)]
 struct CartpoleRecord {
@@ -112,13 +114,9 @@ fn main() -> Result<()> {
         .render_mode(Some("human".to_string()))
         .obs_filter_config(<ObsFilter as GymObsFilter<Obs>>::Config::default())
         .act_filter_config(<ActFilter as GymActFilter<Act>>::Config::default());
-    let mut policy = RandomPolicy;
+    let mut policy = Box::new(RandomPolicy) as _;
 
-    let _ = {
-        let env = Env::build(&env_config, 0)?;
-        Evaluator::new(env, 5)?
-    }
-    .evaluate(&mut policy);
+    let _ = Evaluator::new(&env_config, 42, 5)?.evaluate(&mut policy);
 
     // let mut wtr = csv::WriterBuilder::new()
     //     .has_headers(false)
@@ -140,12 +138,9 @@ fn test_random_cartpole() {
         .name("CartPole-v1".to_string())
         .obs_filter_config(<ObsFilter as GymObsFilter<Obs>>::Config::default())
         .act_filter_config(<ActFilter as GymActFilter<Act>>::Config::default());
-    let mut policy = RandomPolicy;
+    let mut policy = Box::new(RandomPolicy) as _;
 
-    let _ = {
-        let env = Env::build(&env_config, 0).unwrap();
-        Evaluator::new(env, 5)
-    }
-    .unwrap()
-    .evaluate(&mut policy);
+    let _ = Evaluator::new(&env_config, 42, 5)
+        .unwrap()
+        .evaluate(&mut policy);
 }